// src/app/models/ims/inventory-offices-comparative.ts
export class InventoryOfficesComparative {
    branchOffice: string;
    branchOfficeId: number;
    barCode: string;
    productName: string;
    productId: number;
    scaleCode: string;
    area: string;
    inventory: number;
    cost: number;
    sellingFactor: number;
    pvp: number;
    category: string;
    unitNumbers: number;
}